+++ examples/Grids.py
@@ -17,8 +17,7 @@ gbt = {
 
 hwes = {
 	'trend':['add','mul'],
-	'seasonal':['add','mul'],
-	'damped_trend':[True,False]
+	'seasonal':['add','mul']
 }
 
 knn = {
@@ -27,7 +26,18 @@ knn = {
 }
 
 lightgbm = {
-	'max_depth':[-1,2,3]
+	'max_depth':[2,3]
+}
+
+lstm = {
+	'lstm_layer_sizes':[(8,),(8,16,8)],
+	'dropout':[(0,),(0.2,0.2,0)],
+	'activation':['relu','tanh'],
+	'epochs':[5],
+	'batch_size':[32],
+	'random_seed':[20],
+	'shuffle':[True],
+	'verbose':[0]
 }
 
 mlp = {